fix: build errors
Move cert document to different file
Readd old PNG downloading logic
